requirements.txt
jinja2
gitpython